Updated documentation and examples
Added documentation for tree.h and b_search_tree.h.\nUpdated documentation to be more consistent.\nAdded error checking in examples.\nUpdated README.

+++ makefile
@@ -1,3 +1,6 @@
+
+OS := $(shell uname -s 2>/dev/null)
+
 CC := gcc
 CFLAGS := -Wall -Werror -std=c11 -I./include
 SRCDIR := ./src
@@ -7,33 +10,37 @@ BINDIR := ./bin
 
 .PHONY: all mkbin clean
 
-all: e_array_list e_d_linked_list e_graph e_linked_list e_queue e_stack e_hash_map e_tree e_b_search_tree
+all: mkbin e_array_list e_d_linked_list e_graph e_linked_list e_queue e_stack e_hash_map e_tree e_b_search_tree
 
-e_array_list: $(EXAMPLESDIR)/e_array_list.c $(OBJDIR)/array_list.o
+mkbin:
+	mkdir -p $(OBJDIR) 
+	mkdir -p $(BINDIR)
+
+e_array_list: $(EXAMPLESDIR)/e_array_list.c $(OBJDIR)/array_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_d_linked_list: $(EXAMPLESDIR)/e_d_linked_list.c $(OBJDIR)/d_linked_list.o
+e_d_linked_list: $(EXAMPLESDIR)/e_d_linked_list.c $(OBJDIR)/d_linked_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_graph: $(EXAMPLESDIR)/e_graph.c $(OBJDIR)/graph_adj_mat.o $(OBJDIR)/array_list.o
+e_graph: $(EXAMPLESDIR)/e_graph.c $(OBJDIR)/graph_adj_mat.o $(OBJDIR)/array_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_linked_list: $(EXAMPLESDIR)/e_linked_list.c $(OBJDIR)/linked_list.o
+e_linked_list: $(EXAMPLESDIR)/e_linked_list.c $(OBJDIR)/linked_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_queue: $(EXAMPLESDIR)/e_queue.c $(OBJDIR)/queue.o $(OBJDIR)/d_linked_list.o
+e_queue: $(EXAMPLESDIR)/e_queue.c $(OBJDIR)/queue.o $(OBJDIR)/d_linked_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_stack: $(EXAMPLESDIR)/e_stack.c $(OBJDIR)/stack.o $(OBJDIR)/linked_list.o
+e_stack: $(EXAMPLESDIR)/e_stack.c $(OBJDIR)/stack.o $(OBJDIR)/linked_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_hash_map: $(EXAMPLESDIR)/e_hash_map.c $(OBJDIR)/hash_map.o
+e_hash_map: $(EXAMPLESDIR)/e_hash_map.c $(OBJDIR)/hash_map.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
 e_tree: $(EXAMPLESDIR)/e_tree.c $(OBJDIR)/tree.o $(OBJDIR)/array_list.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
-e_b_search_tree: $(EXAMPLESDIR)/e_b_search_tree.c $(OBJDIR)/b_search_tree.o
+e_b_search_tree: $(EXAMPLESDIR)/e_b_search_tree.c $(OBJDIR)/b_search_tree.o $(OBJDIR)/std_dds_utils.o
 	$(CC) $(CFLAGS) -o $(BINDIR)/$@ $^
 
 $(OBJDIR)/array_list.o: $(SRCDIR)/array_list.c
@@ -70,6 +77,10 @@ $(OBJDIR)/std_dds_utils.o: $(SRCDIR)/std_dds_utils.c
 	$(CC) $(CFLAGS) -c -o $@ $^
 
 clean:
+ifeq ($(OS),)
+	if exist $(BINDIR) rmdir /s /q $(BINDIR)
+	if exist $(OBJDIR) rmdir /s /q $(OBJDIR)
+else
 	rm -rf $(BINDIR) $(OBJDIR)
-
+endif
